Extract presenters out of PageLayout repository
The PageLayout class should only be the layouts repository class.
