+++ app/controllers/alchemy/admin/pages_controller.rb
@@ -78,7 +78,7 @@ module Alchemy
 
       def new
         @page ||= Page.new(layoutpage: params[:layoutpage] == "true", parent_id: params[:parent_id])
-        @page_layouts = PageLayout.layouts_for_select(@current_language.id, @page.layoutpage?)
+        @page_layouts = Page.layouts_for_select(@current_language.id, @page.layoutpage?)
         @clipboard = get_clipboard("pages")
         @clipboard_items = Page.all_from_clipboard_for_select(@clipboard, @current_language.id, @page.layoutpage?)
       end
@@ -398,7 +398,7 @@ module Alchemy
       def load_languages_and_layouts
         @language = @current_language
         @languages_with_page_tree = Language.on_current_site.with_root_page
-        @page_layouts = PageLayout.layouts_for_select(@language.id)
+        @page_layouts = Page.layouts_for_select(@language.id)
       end
 
       def set_preview_mode
